migrate analysis watcher to implementation of BaseWatcher that may be extended in enterprise.
